refactor(relays): split relay helpers into modules
