+++ src/lib/relays/config.ts
@@ -1,0 +1,29 @@
+export const RELAYS = {
+  DEFAULT: [
+    'wss://relay.primal.net',
+    'wss://relay.ditto.pub',
+    'wss://nos.lol'
+  ],
+  SEARCH: [
+    'wss://search.nos.today',
+    'wss://relay.ditto.pub',
+    'wss://relay.gathr.gives',
+    'wss://antiprimal.net',
+    'wss://index.hzrd149.com',
+    'wss://nostr.me/relay',
+    'wss://relay.vertexlab.io'
+  ],
+  PROFILE_SEARCH: [
+    'wss://purplepag.es',
+    'wss://search.nos.today',
+    'wss://relay.ditto.pub',
+    'wss://relay.vertexlab.io',
+    'wss://antiprimal.net'
+  ],
+  PREMIUM: [
+    'wss://nostr.wine'
+  ],
+  VERTEX_DVM: [
+    'wss://relay.vertexlab.io'
+  ]
+} as const;
